New fixes

+++ app/controllers/likes_controller.rb
@@ -5,14 +5,18 @@ class LikesController < ApplicationController
   before_action :set_post
 
   def create
-    @post.post_likes.find_or_create_by(user: current_user)
+    @post.likes.find_or_create_by(user: current_user)
     redirect_to posts_url, notice: t('.success')
   end
 
   def destroy
-    like = PostLike.find(params[:id])
-    like.destroy!
-    redirect_to posts_url, notice: t('.success')
+    like = Like.find(params[:id])
+    if like.user == current_user
+      like.destroy!
+      redirect_to posts_url, notice: t('.success')
+    else
+      redirect_to posts_url, alert: t('.failure')
+    end
   end
 
   private
